+++ packages/story-editor/src/components/styleManager/header.js
@@ -32,7 +32,7 @@ import {
 /**
  * Internal dependencies
  */
-import { focusStyle } from '../../../shared';
+import { focusStyle } from '../panels/shared';
 
 const HEADER_HEIGHT = 44;
