+++ packages/story-editor/src/components/styleManager/stories/confirmationDialog.js
