+++ packages/story-editor/src/components/styleManager/styleGroup.js
@@ -24,7 +24,7 @@ import PropTypes from 'prop-types';
 /**
  * Internal dependencies
  */
-import { SAVED_STYLE_HEIGHT, PRESET_TYPES } from '../../../../../constants';
+import { SAVED_STYLE_HEIGHT, PRESET_TYPES } from '../../constants';
 import useKeyboardNavigation from './useKeyboardNavigation';
 import StyleItem from './styleItem';
 
@@ -80,6 +80,7 @@ StyleGroup.propTypes = {
   handleClick: PropTypes.func.isRequired,
   isEditMode: PropTypes.bool,
   buttonWidth: PropTypes.number,
+  activeItemOverlay: PropTypes.node,
 };
 
 export default StyleGroup;
